[SPARK-14872][SQL] Restructure command package
